Rolling back split streaming separation
