+++ chart/templates/worker/split-names/_container.tpl
@@ -2,12 +2,12 @@
 # Copyright 2022 The HuggingFace Authors.
 
 {{- define "containerWorkerSplitNamesStreaming" -}}
-- name: "{{ include "name" . }}-worker-split-names-streaming"
+- name: "{{ include "name" . }}-worker-split-names"
   image: {{ include "services.worker.image" . }}
   imagePullPolicy: {{ .Values.images.pullPolicy }}
   env:
   - name: WORKER_ONLY_JOB_TYPES
-    value: "/split-names-streaming"
+    value: "/split-names"
     # ^ hard-coded
   {{ include "envAssets" . | nindent 2 }}
   {{ include "envCache" . | nindent 2 }}
@@ -16,15 +16,15 @@
   {{ include "envWorker" . | nindent 2 }}
   {{ include "envDatasetsBased" . | nindent 2 }}
   - name: DATASETS_BASED_HF_DATASETS_CACHE
-    value: {{ printf "%s/split-names-streaming/datasets" .Values.cacheDirectory | quote }}
+    value: {{ printf "%s/split-names/datasets" .Values.cacheDirectory | quote }}
   - name: QUEUE_MAX_JOBS_PER_NAMESPACE
     # value: {{ .Values.queue.maxJobsPerNamespace | quote }}
     # overridden
-    value: {{ .Values.splitNamesStreaming.queue.maxJobsPerNamespace | quote }}
+    value: {{ .Values.splitNames.queue.maxJobsPerNamespace | quote }}
   volumeMounts:
   {{ include "volumeMountAssetsRW" . | nindent 2 }}
   {{ include "volumeMountCache" . | nindent 2 }}
   securityContext:
     allowPrivilegeEscalation: false
-  resources: {{ toYaml .Values.splitNamesStreaming.resources | nindent 4 }}
+  resources: {{ toYaml .Values.splitNames.resources | nindent 4 }}
 {{- end -}}
