+++ chart/templates/worker/split-names/deployment.yaml
@@ -4,27 +4,27 @@
 apiVersion: apps/v1
 kind: Deployment
 metadata:
-  labels: {{ include "labels.splitNamesStreaming" . | nindent 4 }}
-  name: "{{ include "name" . }}-worker-split-names-streaming"
+  labels: {{ include "labels.splitNames" . | nindent 4 }}
+  name: "{{ include "name" . }}-worker-split-names"
   namespace: {{ .Release.Namespace }}
 spec:
   progressDeadlineSeconds: 600
-  replicas: {{ .Values.splitNamesStreaming.replicas }}
+  replicas: {{ .Values.splitNames.replicas }}
   revisionHistoryLimit: 10
   selector:
-    matchLabels: {{ include "labels.splitNamesStreaming" . | nindent 6 }}
+    matchLabels: {{ include "labels.splitNames" . | nindent 6 }}
   strategy:
     type: Recreate
   template:
     metadata:
-      labels: {{ include "labels.splitNamesStreaming" . | nindent 8 }}
+      labels: {{ include "labels.splitNames" . | nindent 8 }}
     spec:
       {{- include "image.imagePullSecrets" . | nindent 6 }}
       initContainers:
         {{ include "initContainerAssets" . | nindent 8 }}
         {{ include "initContainerCache" . | nindent 8 }}
-      containers: {{ include "containerWorkerSplitNamesStreaming" . | nindent 8 }}
-      nodeSelector: {{ toYaml .Values.splitNamesStreaming.nodeSelector | nindent 8 }}
-      tolerations: {{ toYaml .Values.splitNamesStreaming.tolerations | nindent 8 }}
+      containers: {{ include "containerWorkerSplitNames" . | nindent 8 }}
+      nodeSelector: {{ toYaml .Values.splitNames.nodeSelector | nindent 8 }}
+      tolerations: {{ toYaml .Values.splitNames.tolerations | nindent 8 }}
       volumes: {{ include "volumeData" . | nindent 8 }}
       securityContext: {{ include "securityContext" . | nindent 8 }}
