+++ libs/libcommon/tests/test_processing_steps.py
@@ -10,7 +10,7 @@ def test_default_graph():
     graph = ProcessingGraph(config.specification)
 
     config_names = graph.get_step("/config-names")
-    split_names_streaming = graph.get_step("/split-names-streaming")
+    split_names = graph.get_step("/split-names")
     splits = graph.get_step("/splits")
     first_rows = graph.get_step("/first-rows")
     parquet_and_dataset_info = graph.get_step("/parquet-and-dataset-info")
@@ -20,13 +20,13 @@ def test_default_graph():
 
     assert config_names is not None
     assert config_names.parent is None
-    assert config_names.children == [split_names_streaming]
+    assert config_names.children == [split_names]
     assert config_names.get_ancestors() == []
 
-    assert split_names_streaming is not None
-    assert split_names_streaming.parent is config_names
-    assert split_names_streaming.children == [first_rows]
-    assert split_names_streaming.get_ancestors() == [config_names]
+    assert split_names is not None
+    assert split_names.parent is config_names
+    assert split_names.children == [first_rows]
+    assert split_names.get_ancestors() == [config_names]
 
     assert splits is not None
     assert splits.parent is None
@@ -34,9 +34,9 @@ def test_default_graph():
     assert splits.get_ancestors() == []
 
     assert first_rows is not None
-    assert first_rows.parent is split_names_streaming
+    assert first_rows.parent is split_names
     assert first_rows.children == []
-    assert first_rows.get_ancestors() == [config_names, split_names_streaming]
+    assert first_rows.get_ancestors() == [config_names, split_names]
 
     assert parquet_and_dataset_info is not None
     assert parquet_and_dataset_info.parent is None
