+++ services/worker/src/worker/job_runners/split_names.py
@@ -64,7 +64,7 @@ def compute_split_names_response(
     hf_token: Optional[str] = None,
 ) -> SplitNamesStreamingResponseContent:
     """
-    Get the response of /split-names-streaming for one specific dataset and config on huggingface.co.
+    Get the response of /split-names for one specific dataset and config on huggingface.co.
     Dataset can be private or gated if you pass an acceptable token.
 
     It is assumed that the dataset exists and can be accessed using the token, and that the config exists in
@@ -73,7 +73,7 @@ def compute_split_names_response(
     This function relies on the streaming mode if the splits are not directly defined in the dataset config. See
     https://github.dev/huggingface/datasets/blob/e183a269067575db8765ee979bd8523d14a1adae/src/datasets/inspect.py#L389-L390
 
-    The /split-names-streaming response generated by this function does not include stats about the split, like
+    The /split-names response generated by this function does not include stats about the split, like
     the size or number of samples. See /dataset-info or /sizes for that.
 
     Args:
@@ -112,7 +112,7 @@ def compute_split_names_response(
 class SplitNamesStreamingJobRunner(DatasetsBasedJobRunner):
     @staticmethod
     def get_job_type() -> str:
-        return "/split-names-streaming"
+        return "/split-names"
 
     @staticmethod
     def get_version() -> str:
